requirements.txt
fastapi==0.105.0
gradio==4.12.0
langchain==0.0.352
open_clip_torch==2.23.0
Pillow==9.3.0
Requests==2.31.0
torch==2.1.2+cpu
uvicorn==0.25.0
transformers==4.36.2
llama_cpp_python==0.2.24